int64maxxing

+++ strukture/Count-min-sketch.go
@@ -9,14 +9,14 @@ import (
 )
 
 type CountMinSketch struct {
-	Width   int
-	Depth   int
-	Greska  float32 // 0 < greska < 1, Sto je manje bolje je
-	Gamma   float32 // 0 < gamma < 1, Sto je vece bolje je
+	Width   int64
+	Depth   int64
+	Greska  float64 // 0 < greska < 1, Sto je manje bolje je
+	Gamma   float64 // 0 < gamma < 1, Sto je vece bolje je
 	Matrica [][]int
 }
 
-func NewCountMinSketch(width, depth int) *CountMinSketch {
+func NewCountMinSketch(width, depth int64) *CountMinSketch {
 	matrica := make([][]int, depth)
 	for i := range matrica {
 		matrica[i] = make([]int, width)
@@ -33,8 +33,8 @@ func deleteCountMiNSketch(sketch *CountMinSketch) {
 }
 
 func (countminsketch *CountMinSketch) Add(item string) {
-	for i := 0; i < countminsketch.Depth; i++ {
-		index := hashfunc.CustomHash(item, countminsketch.Width, i)
+	for i := int64(0); i < countminsketch.Depth; i++ {
+		index := hashfunc.CustomHash(item, int(countminsketch.Width), int(i))
 		countminsketch.Matrica[i][index]++
 	}
 }
@@ -53,8 +53,8 @@ func (countminsketch *CountMinSketch) SerializeCMS() []byte {
 	binary.BigEndian.PutUint64(gammaBytes, math.Float64bits(float64(countminsketch.Gamma)))
 
 	matrixBytes := make([]byte, countminsketch.Width*countminsketch.Depth*8)
-	for i := 0; i < countminsketch.Depth; i++ {
-		for j := 0; j < countminsketch.Width; j++ {
+	for i := int64(0); i < countminsketch.Depth; i++ {
+		for j := int64(0); j < countminsketch.Width; j++ {
 			offset := (i*countminsketch.Width + j) * 8
 			binary.BigEndian.PutUint64(matrixBytes[offset:offset+8], uint64(countminsketch.Matrica[i][j]))
 		}
@@ -87,10 +87,10 @@ func DeserializeCMS(data []byte) (*CountMinSketch, error) {
 	}
 
 	return &CountMinSketch{
-		Width:   width,
-		Depth:   depth,
-		Greska:  float32(greska),
-		Gamma:   float32(gamma),
+		Width:   int64(width),
+		Depth:   int64(depth),
+		Greska:  float64(greska),
+		Gamma:   float64(gamma),
 		Matrica: matrix,
 	}, nil
 }
@@ -98,8 +98,8 @@ func DeserializeCMS(data []byte) (*CountMinSketch, error) {
 func (countminsketch *CountMinSketch) Count(item string) int {
 	var pojave int = math.MaxInt
 
-	for i := 0; i < countminsketch.Depth; i++ {
-		index := hashfunc.CustomHash(item, countminsketch.Width, i)
+	for i := int64(0); i < countminsketch.Depth; i++ {
+		index := hashfunc.CustomHash(item, int(countminsketch.Width), int(i))
 		if countminsketch.Matrica[i][index] < pojave {
 			pojave = countminsketch.Matrica[i][index]
 		}
