+++ strukture/bloomfilter.go
@@ -12,15 +12,15 @@ import (
 
 type BloomFilter struct {
 	BitArray []byte
-	NumHash  int
+	NumHash  int64
 }
 
-func NewBloomFilterWithSize(expectedElements int, falsePositiveRate float64) *BloomFilter {
+func NewBloomFilterWithSize(expectedElements int64, falsePositiveRate float64) *BloomFilter {
 	size := calculateSize(expectedElements, falsePositiveRate)
-	numHash := calculateNumHash(expectedElements, size)
+	numHash := calculateNumHash(expectedElements, int64(size))
 	return &BloomFilter{
 		BitArray: make([]byte, size),
-		NumHash:  numHash,
+		NumHash:  int64(numHash),
 	}
 }
 
@@ -47,7 +47,7 @@ func DeserializeBloomFilter(data []byte) (*BloomFilter, error) {
 	numHash := binary.BigEndian.Uint64(numHashBytes)
 
 	return &BloomFilter{
-		NumHash:  int(numHash),
+		NumHash:  int64(numHash),
 		BitArray: bitArray,
 	}, nil
 }
@@ -57,7 +57,7 @@ func (bf *BloomFilter) Delete() {
 }
 
 func (bf *BloomFilter) Lookup(s string) bool {
-	for i := 0; i < bf.NumHash; i++ {
+	for i := 0; i < int(bf.NumHash); i++ {
 		index := hashfunc.CustomHash(s, len(bf.BitArray), i)
 		if bf.BitArray[index]>>7&1 != 1 {
 			return false
@@ -67,17 +67,17 @@ func (bf *BloomFilter) Lookup(s string) bool {
 }
 
 func (bf *BloomFilter) Insert(s string) {
-	for i := 0; i < bf.NumHash; i++ {
+	for i := 0; i < int(bf.NumHash); i++ {
 		index := hashfunc.CustomHash(s, len(bf.BitArray), i)
 		bf.BitArray[index] |= 1 << 7
 	}
 }
 
-func calculateSize(expectedElements int, falsePositiveRate float64) int {
+func calculateSize(expectedElements int64, falsePositiveRate float64) int {
 	return int(math.Ceil(float64(expectedElements) * math.Abs(math.Log(falsePositiveRate)) / math.Pow(math.Log(2), float64(2))))
 }
 
-func calculateNumHash(expectedElements, size int) int {
+func calculateNumHash(expectedElements, size int64) int {
 	return int(math.Ceil((float64(size) / float64(expectedElements)) * math.Log(2)))
 }
